"dmrg.bib": changed White2002 -> White-can-transform-2002 Added some CG-related papers in "precs_and_solvs.bib" (Van-der-Sluis et al.)
For the future, don't modify the existing names of references. If you don't want to modify all references in your
current papers, just use your own bibs, in the forthcoming texts it would be better to use these common files and
add new papers, if necessary.

For some really new subjects use the file "misc.bib"

Also, if some paper's status changed from e.g. preprint to published, modify it here. In this case, some changes
in year and/or title are possible.